add: implement new CSS styling for page module
A new file `page.module.css` has been added with a plethora of properties defined for various classes to enhance the look of the application. These include classes such as header, slideshow, hero, call to action (cta), and section. The defined properties simplify the process of managing styles on these elements.